pyproject.toml
[project]
name = "cerebras-performance-test"
version = "0.1.0"
tdescription = "Utility for testing Cerebras cloud performance"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.23.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
]
